% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchlisting.R
\name{batch}
\alias{batch}
\title{Batch listing of Argo profiles(Bioargo)}
\usage{
batch(filenames)
}
\arguments{
\item{filenames}{list of Argo profiles(.nc) from the working directory}
}
\description{
Listing of Argo profiles in a batch, it helps in collective analysis of specific sections. Very useful in making time series, contour and filled section contours
}
\examples{

profiles<- list.files()# set working directory where Argo profiles stored

profilelist<- batch(profiles)# A list of Argo profiles



}
\author{
Midhun shah Hussain
}
