% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contour.overlay.R
\name{Contour.overlay}
\alias{Contour.overlay}
\title{Display contours from BioArgo profiles}
\format{The Argument "parameter" character strings can be simply written as small letters could easily work

\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
  \item Backscatter - "backscatter"
  \item Sigma       - "Sigma"
  \item Sigma1000   - "Sigma1000"
}}
\usage{
Contour.overlay(Argolist, parameter, ...)
}
\arguments{
\item{Argolist}{A list of Argo profiles return from function: ExtractBioArgo()}

\item{parameter}{A character string(only work when overview is FALSE)}

\item{...}{All other arguments of basic contour funtion  \code{\link[graphics]{contour}}

\code{\link[graphics]{contour}}}

\item{overview}{A logical value whether you plot the overview, default is FALSE}
}
\description{
This function displays contours from Argo profiles for each parameters. An elaborate overview available for display all major parameters
. Be sure to trim the dataset before plotting(optional)
}
\examples{
Contour.BioArgo(Argolist= myfloatlist,overview = TRUE)# when you want a overview
Contour.BioArgo(Argolist = myfloatlist,parameter = "chlorophyll")# Single paramters
Contour.BioArgo(Argolist = myfloatlist,parameter = "chlorophyll",nlevels=20)# Add more Arguments of contour functions

}
\author{
Midhun shah Hussain
}
