% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_sigma.R
\name{Density_sigmatheta}
\alias{Density_sigmatheta}
\title{Density section graph and density calculations(sigma)}
\usage{
Density_sigmatheta(Argolist, plot_contour = FALSE, pressure.effect = 0,
  ...)
}
\arguments{
\item{Argolist}{A list of Argo profiles return from function: ExtractBioArgo()}

\item{plot_contour}{A logical value whether you plot in contour, default is FALSE}

\item{pressure.effect}{pressure influence on density(Trim profiles according to depths)}

\item{...}{All other arguments for contour plots}
}
\description{
Calculation of density and plotting in section plots.The resources used for this calculations are from Gibbs Sea water(GSW) package.
}
\details{
This function extracts values from each profiles to a section contours. A standalone function doesn't support filledsection contours and overlays



densityat4000<- Density_sigmatheta(test,plot_contour = TRUE,pressure.effect = 4000)
}
