% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractsinglefloat.R
\name{ExtractBioArgo}
\alias{ExtractBioArgo}
\title{A function to Extract data from BioArgo}
\source{
The datasets can be download from following repositories

\href{http://www.argodatamgt.org/Access-to-data/Argo-data-selection}{Argo LAS:For batch downloads}

\href{http://www.ifremer.fr/co-argoFloats/}{Argo: Single downloads}
}
\usage{
ExtractBioArgo(bioarg)
}
\arguments{
\item{bioarg}{= A .nc file from the following sources.}
}
\description{
Extract BioArgo from  raw file in .nc formats.This function is only applicable to BioArgos not for usual Argo floats
}
\author{
Midhun shah Hussain
}
