% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filledsection.R
\name{Filledsectionplots}
\alias{Filledsectionplots}
\title{Section with filled contours and Overlays}
\format{The Argument "parameter" strings can be simply written as small letters could easily work.

\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
}

The Argument "overlay" strings can be simply written as small letters could easily work.
\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
  \item Backscatter - "backscatter"
  \item sigma       - "sigma"
  \item sigma1000   - "sigma1000"
  \item Bouyancy frequency(N2)   - "N2"
}}
\usage{
Filledsectionplots(Argolist, parameter, col.val, overlay = TRUE,
  mld = FALSE, print.mld = FALSE, ...)
}
\arguments{
\item{Argolist}{is a list BioArgo extracted as a output from extractbioArgo function}

\item{parameter}{A character string(only work when overview is FALSE)}

\item{col.val}{amount of colours used in the palette, Here we used matlab.like from library(colorRamps)}

\item{overlay}{Another parameter to compare}

\item{...}{All other arguments of filled.contour \code{\link[graphics]{filled.contour}}}

\item{Parameter}{defines the parameters such as temp, sal,etc}

\item{Argolist}{A list of Argo profiles return from function: ExtractBioArgo()}

\item{overview}{A logical value whether you plot the overview, default is FALSE}

\item{...}{All other arguments of basic contour funtion  \code{\link[graphics]{contour}}

\code{\link[graphics]{contour}}}
}
\description{
This function generates filled contour plots of defined paramters. Before plotting the datasets can be trimmed
 accordingly. This function helps to make comparison plots by make use of line contours as overlays.
}
\examples{

Filledsectionplots(test,parameter = "chlorophyll",col.val = matlab.like(20))#without overlay
Filledsectionplots(test,parameter = "chlorophyll",col.val = matlab.like(20),overlay = "temperature")#with overlay
title(main = "chlorophyll")#Adding a title
legend("bottom",legend = "temperature",lwd = 1)#Adding Legend


Display contours from BioArgo profiles


This function displays contours from Argo profiles for each parameters. An elaborate overview available for display all major parameters
. Be sure to trim the dataset before plotting(optional)
Contour.BioArgo(Argolist= myfloatlist,overview = TRUE)# when you want a overview
Contour.BioArgo(Argolist = myfloatlist,parameter = "chlorophyll")# Single paramters
Contour.BioArgo(Argolist = myfloatlist,parameter = "chlorophyll",nlevels=20)# Add more Arguments of contour functions

}
\author{
Midhun shah Hussain

Midhun shah Hussain
}
