% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_bioArgo.R
\name{Filter_bioArgo}
\alias{Filter_bioArgo}
\title{Filter paramters in BioArgo}
\format{The Argument "parameter" character strings can be simply written as small letters could easily work

\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
  \item Backscatter - "backscatter"
}}
\usage{
Filter_bioArgo(batchlist, parameter, start = NULL, end = NULL,
  month = NULL, year = NULL, day = NULL)
}
\arguments{
\item{batchlist}{A list of Argo profiles return from  \code{\link{batch()}}}

\item{parameter}{A defined number of variables, the formats are shown below}

\item{start}{A value,in which your range starts}

\item{end}{A value,in which your range ends}
}
\description{
This function trims paramters effectively, including latitude, longitude, date and cycle no.Before trims 
find the range using function  \code{\link{bioArgo_Ranger()}}
}
\author{
Midhun shah Hussain
}
