% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_data2ponman}
\alias{get_data2ponman}
\title{Get data from FTP servers of ifremier to your PC}
\usage{
get_data2ponman(mode, location = NULL, year = NULL, month = NULL,
  database = NULL, platform = NULL)
}
\arguments{
\item{mode}{This argument switchs on two different ways to access the datasets

\itemize{
  \item base    - "Access data using database and platform informations"
  \item geotime - "Access data by location, year and month"}}

\item{location}{}

\item{platform}{}
}
\description{
This function enables to access datasets from FTP servers from ifremer.It helps to access databases on two different modes.
}
