% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2squred.R
\name{N2}
\alias{N2}
\title{Brunt–Väisälä frequency calculation}
\usage{
N2(Argolist, plot_contour = FALSE, ...)
}
\arguments{
\item{Argolist}{A list of Argo profiles return from function: ExtractBioArgo()}

\item{plot_contour}{A logical value whether you plot in contour, default is FALSE}

\item{...}{All other arguments for contour plots}
}
\description{
Calculation of BVfrequency and plotting in section plots.The resources used for this calculations are from Gibbs Sea water(GSW) package
}
\examples{


N2(Argolist = test,plot_contour = TRUE)




}
