% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Bioargo.R
\name{plot_BioArgo}
\alias{plot_BioArgo}
\title{A single profile plotting function}
\usage{
plot_BioArgo(x, temperature = FALSE, salinity = FALSE,
  oxygen = FALSE, chlorophyll = FALSE, ...)
}
\arguments{
\item{x}{A single BioArgo file(This argument plots all paramters in a panel)}

\item{temperature?}{Defaults to FALSE}

\item{salinity?}{Defaults to FALSE}

\item{oxygen?}{Defaults to FALSE}

\item{chlorophyll/flouroscence?}{Defaults to FALSE}
}
\value{
A graphical plot of hydrography
}
\description{
plotting overview of graphs, with options to plot single graphs
This function returns all the plots in a single functions
Always be trimmed before plotting according to your needs
}
\examples{
plot_BioArgo(float1)# "float1" return from Extract_bioArgo.
plot_Bioargo(float1,temperature=TRUE)
plot_Bioargo(float1,salinity=TRUE)
}
