% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Argo.R
\name{Sat_bioArgo}
\alias{Sat_bioArgo}
\title{Mapping of BioArgo probes on geospatial  sections}
\usage{
Sat_bioArgo(batchlist, satdata = NULL, satdata.type = NULL,
  lat = NULL, lon = NULL, pix = 0.5, pix.axis = 10,
  legend.month = FALSE, legend.year = FALSE, sat.info = TRUE,
  attr.col = "white", satland = NULL, col.txt = "white", land, ...)
}
\arguments{
\item{batchlist}{list of Argo profiles(BioArgo)}

\item{satdata}{climatology data in .nc format(optional)}

\item{satdata.type}{Type of data 
\itemize{
  \item chlorophyll: "chl"
  \item Sea Surface Temperature: "sst"
  \item Short wave radiation: "swr"}}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{pix}{Pixel density}

\item{pix.axis}{pixel legend}

\item{sat.info}{A logical argument for showing attributes of remote sensing data}

\item{attr.col}{Background colour of attribute box}

\item{satland}{land colour on satellite overlay}

\item{col.txt}{Text color of attribute box}

\item{land}{color argument for land, when remote sensing data are absent}

\item{...}{Possible arguments for basic or raster maping}

\item{Legend.month}{Legend for months}

\item{Legend.year}{Legend for years}
}
\description{
Mapping of BioArgo Trajectory to get an overiview on its spatial coverage. Along with this, a climatology implimentation using remote sensing data-sets can be possible.This function can be efficiently used by adding manual coordinates.
This mapping system also helps in monthly distribution of BioArgo floats
}
\examples{
plot_locations_only<-  Sat_bioArgo(profile2013,lat = c(15,22),lon = c(65,70),land = "red")
plot_with_rsdata<- Sat_bioArgo(profile2013,lat = c(15,22),lon = c(65,70),
        satdata = "~/MEGAsync/Data/midhunshah-Argo/A20030602018090.L3m_MC_CHL_chlor_a_4km.nc",satdata.type = 'chl')
plot_rsdata_Legend<- Sat_bioArgo(profile2013,lat = c(15,22),lon = c(65,70),
         satdata = "~/MEGAsync/Data/midhunshah-Argo/A20030602018090.L3m_MC_CHL_chlor_a_4km.nc",satdata.type = 'chl',legend.month=TRUE)       
        
        
}
\author{
Midhun shah Hussain
}
