% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filledsection.R
\name{Filledsectionplots}
\alias{Filledsectionplots}
\title{Section with filled contours and Overlays}
\format{The Argument "parameter" strings can be simply written as small letters could easily work.

\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
}

The Argument "overlay" strings can be simply written as small letters could easily work.
\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
  \item Backscatter - "backscatter"
  \item sigma       - "sigma"
  \item sigma1000   - "sigma1000"
}}
\usage{
Filledsectionplots(Argolist, parameter, col.val, overlay = TRUE,
  mld = FALSE, print.mld = FALSE, ...)
}
\arguments{
\item{Argolist}{is a list BioArgo extracted as a output from extractbioArgo function}

\item{col.val}{amount of colours used in the palette, Here we used matlab.like from library(colorRamps)}

\item{overlay}{Another parameter to compare}

\item{...}{All other arguments of filled.contour \code{\link[graphics]{filled.contour}}}

\item{Parameter}{defines the parameters such as temp, sal,etc}
}
\description{
This function generates filled contour plots of defined paramters. Before plotting the datasets can be trimmed
 accordingly. This function helps to make comparison plots by make use of line contours as overlays.
}
\examples{

Filledsectionplots(test,parameter = "chlorophyll",col.val = matlab.like(20))#without overlay
Filledsectionplots(test,parameter = "chlorophyll",col.val = matlab.like(20),overlay = "temperature")#with overlay
title(main = "chlorophyll")#Adding a title
legend("bottom",legend = "temperature",lwd = 1)#Adding Legend


}
\author{
Midhun shah Hussain
}
