% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_bioArgo.R
\name{Filter_bioArgo}
\alias{Filter_bioArgo}
\title{Filter paramters in BioArgo f}
\format{The Argument "parameter" character strings can be simply written as small letters could easily work

\itemize{
  \item Temperature - "temperature"
  \item Salinity    - "salinity"
  \item Oxygen      - "oxygen"
  \item Chlorophyll - "chlorophyll"
  \item Backscatter - "backscatter"
}
Parameter "date" formating for months
\itemize{
  \item January  -"01"
  \item February -"02"
  \item March    -"03"
  \item . . . ..
  \item December -"12"
}}
\usage{
Filter_bioArgo(batchlist, parameter, start = NULL, end = NULL,
  month = NULL, year = NULL, day = NULL)
}
\arguments{
\item{batchlist}{A list of Argo profiles return from  \code{\link{batch()}}}

\item{parameter}{A defined number of variables, the formats are shown below}

\item{start}{A value,in which your range starts, Not valued when parameter is "date"}

\item{end}{A value,in which your range ends,Not valued when parameter is "date"}

\item{month}{A value as "character", when parameter is "date"}

\item{year}{A value as "character", when parameter is "date"}

\item{day}{A value as "character", when parameter is "date"}
}
\description{
This function trims paramters effectively, including latitude, longitude, date and cycle no.Before trims
find the range using function  \code{\link{bioArgo_Ranger()}}
}
\examples{
Argo_February<- Filter_bioArgo(test,parameter = "date",month = "02")#select months
Argo_2013<- Filter_bioArgo(test,parameter = "date",year = "2013")#select years
depth20.30<- Filter_bioArgo(test,parameter = "pressure",start = 20,end = 30)# Be careful that the values are numeric




}
\author{
Midhun shah Hussain
}
