% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data2ponman.R
\name{get_data2ponman}
\alias{get_data2ponman}
\title{Get data from FTP servers of ifremier to your PC}
\usage{
get_data2ponman(mode, location = NULL, year = NULL, month = NULL,
  database = NULL, platform = NULL)
}
\arguments{
\item{mode}{This argument switchs on two different ways to access the datasets

\itemize{
  \item base    - "Access data using database and platform informations"
  \item geotime - "Access data by location, year and month"}}

\item{location}{Three locations are currently in operation

\itemize{
  \item Indian Ocean: "indian_ocean"
  \item Atlantic Ocean: "atlantic_ocean"
  \item Pacific Ocean: "pacific_ocean"}}

\item{year}{datasets are found to be available from 1997 to 2018.}

\item{month}{Months are availble on the regular format of "01" as January and "12" as December}

\item{database}{The database options 
\itemize{
 \item "aoml"
 \item "bodc"        
 \item "coriolis"
 \item "csio"
 \item "csiro"        
 \item "incois"        
 \item "jma"        
 \item "kma"        
 \item "kordi"    
 \item "meds"    
 \item "nmdis"    }}

\item{platform}{which is a number(eg:"2902092")  need to know before you}
}
\description{
This function enables to access datasets from FTP servers from ifremer.It helps to access databases on two different modes.
}
\examples{
extractargpdata<- get_data2ponman(mode = "base",database = "incois",platform = "2902092")
extractlocation_time<- get_data2ponman(mode = "geotime",location = "indian_ocean",year = "2010",month = "02")

}
