% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Argo.R
\name{Sat_bioArgo}
\alias{Sat_bioArgo}
\title{Mapping of BioArgo probes on geospatial  sections}
\usage{
Sat_bioArgo(batchlist, satdata = NULL, lat = NULL, lon = NULL,
  legend.month = FALSE, legend.year = FALSE, ...)
}
\arguments{
\item{batchlist}{list of Argo profiles(BioArgo)}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{...}{Possible arguments for basic or raster maping}

\item{satadata}{climatology data in .nc format(optional)}

\item{Legend.month}{Legend for months}

\item{Legend.year}{Legend for years}
}
\description{
Mapping of BioArgo Trajectory to get an overiview on its spatial coverage. Along with this, a climatology implimentation using remote sensing data-sets can be possible.This function can be efficiently used by adding manual coordinates.
This mapping system also helps in monthly distribution of BioArgo floats
}
